data.sql favorites not yet working

+++ src/main/java/com/graduation/backend/service/FavoritesServiceImpl.java
@@ -3,6 +3,8 @@ package com.graduation.backend.service;
 import com.graduation.backend.dto.FavoritesDto;
 import com.graduation.backend.exceptions.RecordNotFoundException;
 import com.graduation.backend.model.Favorites;
+import com.graduation.backend.model.Travels;
+import com.graduation.backend.model.Users;
 import com.graduation.backend.repository.FavoritesRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -44,20 +46,29 @@ public class FavoritesServiceImpl implements FavoritesService {
     public Favorites createFavorites(FavoritesDto favo) {
         Favorites f = new Favorites();
         f.setId(favo.getId());
-
-//        f.setTravels(favo.getTravels());
-//        f.setUsers(favo.getUsers());
-
+        f.setTravels((Travels) favo.getTravels());
+        f.setUsers((Users) favo.getUsers());
         return this.repos.save(f);
     }
 
-
     @Override
-    public String deleteFavorites(Long id) {
-        Favorites favorite = repos.findById(id).orElseThrow(() -> new RecordNotFoundException("Favorite not found"));
-        repos.deleteById(id);
-        return "Favorite deleted";
-
+    public FavoritesDto deleteFavorites(Long id) {
+        Optional<Favorites> favorites = repos.findById(id);
+        if(favorites.isPresent()){
+            FavoritesDto favoritesDto = mapper.map(favorites.get(), FavoritesDto.class);
+            repos.deleteById(id);
+            return favoritesDto;
+        } else {
+            throw new RecordNotFoundException("Unable to delete favorite");
+        }
     }
+
+//    @Override
+//    public String deleteFavorites(Long id) {
+//        Favorites favorite = repos.findById(id).orElseThrow(() -> new RecordNotFoundException("Favorite not found"));
+//        repos.deleteById(id);
+//        return "Favorite deleted";
+//
+//    }
 
 }
